speech-to-text-app/requirements.txt
flask>=1.0
numpy<2
torch
openai-whisper
ffmpeg
ffmpeg-python